backend-ms1/src/test/java/ma/zyn/app/unit/service/impl/instructor/course/LanguageInstructorServiceImplTest.java
package ma.zyn.app.unit.service.impl.instructor.course;

import ma.zyn.app.bean.core.course.Language;
import ma.zyn.app.dao.facade.core.course.LanguageDao;
import ma.zyn.app.service.impl.instructor.course.LanguageInstructorServiceImpl;

import java.util.*;
import java.math.BigDecimal;
import java.time.LocalDateTime;
import java.util.stream.Collectors;
import java.util.stream.IntStream;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Disabled;
import org.junit.jupiter.api.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import java.time.LocalDateTime;



import org.springframework.boot.test.context.SpringBootTest;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;


@SpringBootTest
class LanguageInstructorServiceImplTest {

    @Mock
    private LanguageDao repository;
    private AutoCloseable autoCloseable;
    private LanguageInstructorServiceImpl underTest;



    @BeforeEach
    void setUp() {
        autoCloseable = MockitoAnnotations.openMocks(this);
        underTest = new LanguageInstructorServiceImpl(repository);
    }

    @AfterEach
    void tearDown() throws Exception {
        autoCloseable.close();
    }

    @Test
    void canGetAllLanguage() {
         //when
        underTest.findAll();
        verify(repository).findAll();
    }

    @Test
    void itShouldSaveLanguage() {
        // Given
        Language toSave = constructSample(1);
        when(repository.save(toSave)).thenReturn(toSave);

        // When
        underTest.create(toSave);

        // Then
        verify(repository).save(toSave);
    }

    @Test
    void itShouldDeleteLanguage() {
        // Given
        Long idToDelete = 1L;
        when(repository.existsById(idToDelete)).thenReturn(true);

        // When
        underTest.deleteById(idToDelete);

        // Then
        verify(repository).deleteById(idToDelete);
    }
    @Test
    void itShouldGetLanguageById() {
        // Given
        Long idToRetrieve = 1L; // Example Language ID to retrieve
        Language expected = new Language(); // You need to replace Language with your actual class
        expected.setId(idToRetrieve);
        when(repository.findById(idToRetrieve)).thenReturn(java.util.Optional.of(expected));

        // When
        Language result = underTest.findById(idToRetrieve);

        // Then
        assertEquals(expected, result);
    }
	
	private Language constructSample(int i) {
		Language given = new Language();
        given.setCode("code-"+i);
        given.setLabel("label-"+i);
        given.setStyle("style-"+i);
        return given;
    }

}
